requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.25.2
aiohttp==3.9.1
click==8.1.7
alembic==1.13.1
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.12.0
ruff==0.1.8
mypy==1.7.1

# Phase 1: Authentication & Email
PyJWT==2.8.1
python-multipart==0.0.6
slowapi==0.1.9
bcrypt==4.1.1
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.106.0
msgraph-core==1.0.0
azure-identity==1.14.0

# Phase 1: AI Email Generation
openai==1.9.0
tenacity==8.2.3

# Phase 1: Reply Parsing
textblob==0.17.1

# Phase 1: Batch Operations & Jobs
celery==5.3.4
redis==5.0.1

# Phase 2: Analytics & Webhooks
sqlalchemy-utils==0.41.1
pydantic-extra-types==2.4.1

